# #5592: Optimize Falcon 7b lm head matmul