--- a/models/demos/falcon7b/tt/falcon_lm_head.py
+++ b/models/demos/falcon7b/tt/falcon_lm_head.py
@@ -1,3 +1,7 @@
+# SPDX-FileCopyrightText: © 2023 Tenstorrent Inc.
+
+# SPDX-License-Identifier: Apache-2.0
+
 import torch
 from typing import List