Add a function for Falcon 7b lm head with seq len > 512. The function splits matmul into 4 or 8 2d matmuls, and expects the weights to be sliced accordingly. 

For 2k seq len, kernel time went from 400ms to 17ms (total kernel time for all slices and concats). For 1k seq len kernel time went from 200ms to 9ms. Other 32-multiple sequence lengths between 512 and 2048 are also supported.

The function is currently used only in the test - the idea is to incorporate it to the model with other pending optimizations; then some more optimizations and refactoring will be done (eg padding tensor will probably be reused on multiple spots, model config will be used instead of some fixed constants like grid, subblock dimensions etc.).